creating, editing and deleting User now triggers websock notification event correctly
adjustment for cases where info didn't change (will scale...)

nested 'method' after 'model' on notificationRoutines for future scaling

added 'hasAvatar' prop to dynamicDataTableCell to render things correctly

+++ frontend/src/forms/delete/DeleteFormModel.jsx
@@ -31,7 +31,7 @@ const DeleteFormModel = ({
 }) => {
   const handleDelete = async () => {
     try {
-      const res = await api.delete(`/delete/${model}/${selectedItem._id}`);
+      const res = await api.delete(`/delete/${userId}/${model}/${selectedItem._id}`);
       if (res.data) {
         toast.success(`${selectedItem.name} Deletado!`, {
           closeOnClick: true,
